requirements.txt

onnxruntime-gpu
matplotlib
numpy
tqdm
torchvision
opencv-python-headless
av